Source/Python/cf_framework/hsp_cf_framework/requirements.txt
hvac
cloudfoundry-client
Jinja2
requests
urllib3
PyYAML
future
mock
jmespath
datetime
pytz
jsonpath_rw_ext
pathlib
tabulate
